Client latency reduction
Altered a ton of the logic that controls the construction and usage of
control packets in order to reduce input lag for the client(s) and make
the host input lag similar to that of the client. The client inputs no
longer have to perform a two-way trip in order to be used, and merely go
in one direction, to the host.